separate search components
