+++ src/pages/search/search.jsx
@@ -1,48 +1,13 @@
-import { useRef, useState } from 'react'
 import styles from './search.module.css'
+import SearchHeader from './search_header/search_header';
+import SearchInfo from './search_info/search_info';
 
-export default function Search({cards}) {
-
-  const inputRef = useRef();
-
-  const handleSearch = () => {
-    const value = inputRef.current.value;
-    // onSearch(value)
-    console.log(value)
-  }
-
-  const onClick = () => {
-    handleSearch();
-  }
-
-  const onKeyPress = (event) => {
-    if (event.key === 'Enter') {
-      handleSearch();
-    }
-  }
+export default function Search({cards, onSearch}) {
 
   return (
-    <div className={styles.search}>
-      <div className={styles.logo_container}>
-        <h2 className={styles.title}>Search by Name</h2>
-      </div>
-      <input ref={inputRef} className={styles.input} type="search" placeholder="Search..." onKeyPress={onKeyPress}/>
-      <button className={styles.button} type="submit" onClick={onClick}>Search</button>
-      <div className={styles.card_container}>
-      </div>
-    </div>
+    <section className={styles.search}>
+      <SearchHeader onSearch={onSearch}/>
+      <SearchInfo cards={cards} />
+    </section>
   )
-}
-
-function getStyles(theme) {
-  switch (theme) {
-    case 'dark':
-      return styles.dark;
-    case 'light':
-      return styles.light;
-    case 'colorful':
-      return styles.colorful;
-    default:
-      throw new Error(`unknown theme: ${theme}`);
-  }
 }
