+++ src/pages/search/search_header/search_header.jsx
@@ -1,0 +1,31 @@
+import styles from './search_header.module.css'
+import { useRef } from 'react'
+
+export default function SearchHeader({onSearch}) {
+  const inputRef = useRef();
+
+  const handleSearch = () => {
+    const value = inputRef.current.value;
+    onSearch(value)
+  }
+
+  const onClick = () => {
+    handleSearch();
+  }
+
+  const onKeyPress = (event) => {
+    if (event.key === 'Enter') {
+      handleSearch();
+    }
+  }
+
+  return (
+    <header className={styles.search}>
+      <div className={styles.logo_container}>
+        <h2 className={styles.title}>Search by Name</h2>
+      </div>
+      <input ref={inputRef} className={styles.input} type="search" placeholder="Search..." onKeyPress={onKeyPress}/>
+      <button className={styles.button} type="submit" onClick={onClick}>Search</button>
+    </header>
+  )
+}
